CRUD_DJANGO_REACT/requirements.txt
asgiref==3.8.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.8
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
filelock==3.14.0
Flask==3.0.2
idna==3.6
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
pillow==10.3.0
platformdirs==4.2.1
requests==2.31.0
setuptools==69.1.1
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.26.1
Werkzeug==3.0.1
